python/requirements.txt
# SharePoint and authentication dependencies
Office365-REST-Python-Client>=2.5.0
msal>=1.24.0

# Utility dependencies
argparse>=1.4.0
requests>=2.31.0
urllib3>=1.26.0